Move region into shared directory
