+++ src/exporters/mod.rs
@@ -11,3 +11,4 @@ pub mod xray;
 
 pub(crate) mod crypto_init_tests;
 pub(crate) mod http;
+pub(crate) mod shared;
